refactor(interpreter): move helper modules into utils
