+++ crates/ltmf/src/interpreter/text/color.rs
@@ -1,6 +1,7 @@
 use serde_json::Value;
 
-use crate::interpreter::{get_marks::get_marks_by_type, rgba_to_hex::color_to_wikidot_hex};
+use crate::interpreter::utils::rgba_to_hex::color_to_wikidot_hex;
+use crate::interpreter::utils::get_marks::get_marks_by_type;
 
 pub fn interpret_color_text(node: &Value, output: String) -> Result<String, String> {
     let Some(color) = color_text_mark(node) else {
