+++ crates/ltmf/src/interpreter/text/underline.rs
@@ -1,6 +1,6 @@
 use serde_json::Value;
 
-use crate::interpreter::get_marks::has_mark;
+use crate::interpreter::utils::get_marks::has_mark;
 
 pub fn interpret_underline_text(node: &Value, output: String) -> Result<String, String> {
     if !has_mark(node, "underline") {
