+++ crates/ltmf/src/interpreter/utils/rgba_to_hex.rs
